+++ CHANGES.rst
@@ -29,6 +29,7 @@ New features:
 - compute_mean_max uses vmpy.algorithms.power_duration_curve
 - compute_weighted_average_power uses vmpy.metrics.normalized_power
 - compute_power_per_kg uses vmpy.metrics.wpk
+- compute_power_zones based on athlete.ftp, custom ftp or custom zones
 
 Bug fixes: